+++ oConsole.py
@@ -1,10 +1,35 @@
 import codecs, ctypes, ctypes.wintypes, locale, sys, threading;
 #oStdOut = codecs.getwriter(locale.getpreferredencoding())(sys.stdout, "replace");
 
+SHORT = ctypes.c_short;
+WORD = ctypes.c_ushort;
 STD_INPUT_HANDLE = -10
 STD_OUTPUT_HANDLE = -11
 STD_ERROR_HANDLE = -12
 NORMAL_COLOR = 0x7;
+
+class COORD(ctypes.Structure):
+  _fields_ = [
+    ("X", SHORT),
+    ("Y", SHORT)
+  ];
+
+class SMALL_RECT(ctypes.Structure):
+  _fields_ = [
+    ("Left", SHORT),
+    ("Top", SHORT),
+    ("Right", SHORT),
+    ("Bottom", SHORT)
+  ];
+    
+class CONSOLE_SCREEN_BUFFER_INFO(ctypes.Structure):
+  _fields_ = [
+    ("dwSize", COORD),
+    ("dwCursorPosition", COORD),
+    ("wAttributes", WORD),
+    ("srWindow", SMALL_RECT),
+    ("dwMaximumWindowSize", COORD)
+  ];
 
 # Non-unicode strings are assumed to be CP437. We have an indexed table to
 # convert CP437 to unicode (index range 0-255 => unicode char) and a dict to
@@ -63,12 +88,43 @@ class cConsole(object):
     dwMode = ctypes.wintypes.DWORD(0);
     oConsole.bStdOutIsConsole = ctypes.windll.kernel32.GetConsoleMode(oConsole.hStdOut, ctypes.byref(dwMode));
     oConsole.bByteOrderMarkWritten = False;
+    oConsole.uDefaultColor = -1;
+  
+  def __foGetConsoleScreenBufferInfo(oConsole):
+    assert oConsole.bStdOutIsConsole, \
+        "Cannot set colors when output is redirected";
+    oConsoleScreenBufferInfo = CONSOLE_SCREEN_BUFFER_INFO()
+    assert ctypes.windll.kernel32.GetConsoleScreenBufferInfo(oConsole.hStdOut, ctypes.byref(oConsoleScreenBufferInfo)), \
+        "GetConsoleScreenBufferInfo(%d, ...) => Error %08X" % \
+        (oConsole.hStdOut, ctypes.windll.kernel32.GetLastError());
+    return oConsoleScreenBufferInfo;
+  
+  @property
+  def uCurrentColor(oConsole):
+    oConsoleScreenBufferInfo = oConsole.__foGetConsoleScreenBufferInfo();
+    return oConsoleScreenBufferInfo.wAttributes & 0xFF;
 
+  @property
+  def uWindowWidth(oConsole):
+    oConsoleScreenBufferInfo = oConsole.__foGetConsoleScreenBufferInfo();
+    return oConsoleScreenBufferInfo.srWindow.Right - oConsoleScreenBufferInfo.srWindow.Left;
+  
+  @property
+  def uWidth(oConsole):
+    oConsoleScreenBufferInfo = oConsole.__foGetConsoleScreenBufferInfo();
+    return oConsoleScreenBufferInfo.dwSize.X;
+  
   def __fSetColor(oConsole, uColor):
     assert oConsole.bStdOutIsConsole, \
         "Cannot set colors when output is redirected";
-    assert uColor in xrange(0, 0x100), \
+
+    uMask = uColor >> 8;
+    assert uMask in xrange(0, 0x100), \
         "You cannot use color 0x%X; maybe you are trying to print a number without converting it to a string?" % uColor;
+    uColor &= 0xFF;
+    if uMask:
+      uCurrentColor = oConsole.uCurrentColor;
+      uColor = (uCurrentColor & (uMask ^ 0xFF)) + (uColor & uMask);
     assert ctypes.windll.kernel32.SetConsoleTextAttribute(oConsole.hStdOut, uColor), \
         "SetConsoleTextAttribute(%d, %d) => Error %08X" % \
         (oConsole.hStdOut, uColor, ctypes.windll.kernel32.GetLastError());
@@ -91,47 +147,73 @@ class cConsole(object):
           (sWriteFunctionName, oConsole.hStdOut, uCharsToWrite, ctypes.windll.kernel32.GetLastError());
       sMessage = sMessage[dwCharsWritten.value:];
 
-  def __fOutputHelper(oConsole, axCharsAndColors, bNewLine):
-    global goLock, guLastLineLength;
+  def __fOutputHelper(oConsole, axCharsAndColors, bIsStatusMessage):
+    assert oConsole.bStdOutIsConsole or not bIsStatusMessage, \
+        "Status messages should not be output when output is redirected.";
     oConsole.oLock.acquire();
     try:
       # Go to the start of the current line if needed
       if oConsole.uLastLineLength:
         oConsole.__fWriteOutput(oConsole.bStdOutIsConsole and u"\r" or "\r");
+      uCharsOutput = 0;
+      # setup colors if outputting to a console.
       bColorWasSet = False;
-      uNewLineLength = 0;
+      if oConsole.bStdOutIsConsole:
+        uColumns = oConsole.uWidth;
+        uOriginalColor = oConsole.uCurrentColor;
+        if oConsole.uDefaultColor != -1:
+          oConsole.__fSetColor(oConsole.uDefaultColor);
+          bColorWasSet = True;
       try:
         for xCharsOrColor in axCharsAndColors:
           if isinstance(xCharsOrColor, int) or isinstance(xCharsOrColor, long):
             if oConsole.bStdOutIsConsole: # If output is redirected, colors will not be set, so don't try
-              if xCharsOrColor == -1: xCharsOrColor = NORMAL_COLOR;
+              if xCharsOrColor == -1: xCharsOrColor = uOriginalColor;
               oConsole.__fSetColor(xCharsOrColor);
               bColorWasSet = True;
           else:
             assert isinstance(xCharsOrColor, str) or isinstance(xCharsOrColor, unicode), \
                 "You cannot print %s (type = %s) directly; it must be converted to a string" % (repr(xCharsOrColor), xCharsOrColor.__class__.__name__);
-            uNewLineLength += len(xCharsOrColor);
+            if bIsStatusMessage and uCharsOutput + len(xCharsOrColor) >= uColumns:
+              # Do not let a status message span multiple lines.
+              xCharsOrColor = xCharsOrColor[:uColumns - uCharsOutput - 1];
             oConsole.__fWriteOutput(xCharsOrColor);
+            uCharsOutput += len(xCharsOrColor);
+            if bIsStatusMessage and uCharsOutput == uColumns - 1:
+              # We've reached the end of the line and the remained of the arguments will not be output.
+              break;
       finally:
         if bColorWasSet:
-          oConsole.__fSetColor(NORMAL_COLOR);
-      if uNewLineLength < oConsole.uLastLineLength:
-        oConsole.__fWriteOutput(u" " * (oConsole.uLastLineLength - uNewLineLength));
-      if bNewLine:
-        oConsole.__fWriteOutput(oConsole.bStdOutIsConsole and u"\r\n" or "\r\n");
-        oConsole.uLastLineLength = 0;
+          oConsole.__fSetColor(uOriginalColor);
+      if oConsole.bStdOutIsConsole:
+        # Optionally output some padding if this is a status message that is smaller than the previous status message.
+        # Then go back to the start of the line and move to the next line if this is not a status message.
+        oConsole.__fWriteOutput("".join([
+          uCharsOutput < oConsole.uLastLineLength and u" " * (oConsole.uLastLineLength - uCharsOutput) or "",
+          bIsStatusMessage and u"\r" or u"\r\n",
+        ]));
+        oConsole.uLastLineLength = bIsStatusMessage and uCharsOutput or 0;
       else:
-        oConsole.__fWriteOutput(oConsole.bStdOutIsConsole and u"\r" or "\r");
-        oConsole.uLastLineLength = uNewLineLength;
+        oConsole.__fWriteOutput("\n");
+        oConsole.uLastLineLength = 0;
     finally:
       oConsole.oLock.release();
 
   def fPrint(oConsole, *axCharsAndColors):
-    oConsole.__fOutputHelper(axCharsAndColors, True);
+    oConsole.__fOutputHelper(axCharsAndColors, False);
 
   def fStatus(oConsole, *axCharsAndColors):
     # If output is redirected, do not output status messages
     if oConsole.bStdOutIsConsole:
-      oConsole.__fOutputHelper(axCharsAndColors, False);
+      oConsole.__fOutputHelper(axCharsAndColors, True);
+  
+  def fProgressBar(oConsole, nProgress, sMessage = "", uProgressColor = 0xA0, uBarColor = 0x2A):
+    assert nProgress >=0 and nProgress <= 1, \
+        "Progress must be [0, 1], not %s" % nProgress;
+    if oConsole.bStdOutIsConsole:
+      uBarWidth = oConsole.uWindowWidth - 1;
+      sBar = sMessage.center(uBarWidth);
+      uProgress = long(oConsole.uWindowWidth * nProgress);
+      oConsole.__fOutputHelper([uProgressColor, sBar[:uProgress], uBarColor, sBar[uProgress:]], True);
 
 oConsole = cConsole();